spark-server/requirements.txt
# Spark AI Server Dependencies

# WebSocket server
websockets>=12.0

# HTTP client/server
aiohttp>=3.9.0

# Audio processing
edge-tts>=6.1.0

# Optional: opus encoding/decoding (requires system libopus)
# opuslib>=3.0.1

# Utilities
python-dotenv>=1.0.0

# For production deployment
gunicorn>=21.0.0
uvloop>=0.19.0
